chore: Move more expressions from core crate to spark-expr crate (#1152)
* move aggregate expressions to spark-expr crate

* move more expressions

* move benchmark

* normalize_nan

* bitwise not

* comet scalar funcs

* update bench imports
