+++ native/spark-expr/src/avg.rs
